requirements.txt
aiogram==3.4.1
aiosqlite==0.20.0
requests==2.32.3
python-dotenv==1.0.1